Entity-Matching/mpc-cdf-toolkit/modules/data_entity_matching/functions/fn_standard_entity_matching/requirements.txt
pydantic
cognite-sdk
pandas
annotated-types==0.7.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cognite-sdk==7.75.0
cryptography==44.0.3
dotenv==0.9.9
idna==3.10
msal==1.32.3
oauthlib==3.2.2
packaging==25.0
protobuf==6.30.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
croniter